Added new interfaces
