--- YUP/YUP.App/Contracts/YupManger.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using YUP.App.Services;
-
-namespace YUP.App.Contracts
-{
-    public class YupManger : IYupiManager
-    {
-        public string test()
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
