+++ YUP/YUP.App/Contracts/YupRepository.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using YUP.App.Services;
+
+namespace YUP.App.Contracts
+{
+    public class YupManger : IYupiManager
+    {
+        public string test()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
