+++ YUP/YUP.App/Enums.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace YUP.App
+{
+    /// <summary>
+    /// Enum defining in what mode are we running ( based on internet access)
+    /// </summary>
+    public enum YupMode
+    {
+        Online,
+        Offline
+    }
+}
