+++ YUP/YUP.App/Events/PlayerStateChangedArgs.cs
@@ -5,7 +5,7 @@ namespace YUP.App.Events
 {
     public class PlayerStateChangedArgs : EventArgs
     {
-        public MediaPlaybackState   PlaybackState   { get; set; }
-        public string               VideoId         { get; set; }
+        public MediaPlaybackState   playbackState   { get; set; }
+        public string               videoId         { get; set; }
     }
 }
