+++ YUP/YUP.App/Services/IYupRepository.cs
@@ -1,0 +1,7 @@
+﻿namespace YUP.App.Services
+{
+    public interface IYupiManager
+    {
+        string test();
+    }
+}
