+++ YUP/YUP.App/Services/IYupSettings.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace YUP.App.Services
+{
+    /// <summary>
+    /// interface used for managing settings in application
+    /// </summary>
+    public interface IYupSettings
+    {
+    }
+}
