--- YUP/YUP.App/Services/IYupiManager.cs
@@ -1,7 +1,0 @@
-﻿namespace YUP.App.Services
-{
-    public interface IYupiManager
-    {
-        string test();
-    }
-}
